Create CSS code for making website responsiver
This file contains CSS code which will make the website responsive. Responsive web design (RWD) or responsive design is an approach to web design that aims to make web pages render well on a variety of devices and window or screen sizes from minimum to maximum display size to ensure usability and satisfaction.